o.c.opibuilder: merge identical OpenDisplayTarget enums.
